--- packages/sidecar/src/launch-editor.d.ts
@@ -1,1 +1,0 @@
-declare module "launch-editor";
