+++ packages/sidecar/src/routes/clear.ts
@@ -1,0 +1,11 @@
+import { Hono } from "hono";
+import { getBuffer } from "~/utils/index.js";
+
+const router = new Hono();
+
+router.delete("/", ctx => {
+  getBuffer().clear();
+  return ctx.text("Cleared");
+});
+
+export default router;
