+++ packages/sidecar/src/routes/health.ts
@@ -1,0 +1,7 @@
+import { Hono } from "hono";
+
+const router = new Hono();
+
+router.get("/", ctx => ctx.text("OK"));
+
+export default router;
