+++ packages/sidecar/src/routes/mcp/__tests__/test_envelopes.ts
