+++ packages/sidecar/src/routes/mcp/__tests__/tools.test.ts
