+++ packages/sidecar/src/routes/mcp/__tests__/traces.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it } from "vitest";
-import { EventContainer } from "../../eventContainer.js";
+import { EventContainer } from "../../../utils/eventContainer.js";
 import { buildSpanTree, extractTracesFromEnvelopes, formatTraceSummary, renderSpanTree } from "../utils/traces.js";
 import { envelopeSecondTransactionEvent, envelopeTransactionEvent } from "./test_envelopes.js";
