+++ packages/sidecar/src/routes/mcp/constants.ts
