+++ packages/sidecar/src/routes/mcp/formatting.ts
