+++ packages/sidecar/src/routes/mcp/parsing.ts
