+++ packages/sidecar/src/routes/mcp/schema.ts
