+++ packages/sidecar/src/routes/mcp/types.ts
