+++ packages/sidecar/src/routes/mcp/utils/index.ts
