+++ packages/sidecar/src/types/utils.ts
@@ -1,0 +1,48 @@
+export type IncomingPayloadCallback = (body: string) => void;
+
+export type SidecarLogger = {
+  info: (message: string) => void;
+  warn: (message: string) => void;
+  error: (message: string) => void;
+  debug: (message: string) => void;
+};
+
+export type SideCarOptions = {
+  /**
+   * The port on which the sidecar should listen.
+   * Defaults to 8969.
+   */
+  port?: string | number;
+
+  /**
+   * A logger that implements the SidecarLogger interface.
+   * Use this to inject your custom logger implementation.
+   *
+   * @default - a simple logger logging to the console.
+   */
+  logger?: SidecarLogger;
+
+  /**
+   * The base path from where the static files should be served.
+   */
+  basePath?: string;
+
+  filesToServe?: Record<string, Buffer>;
+
+  /**
+   * More verbose logging.
+   */
+  debug?: boolean;
+
+  /**
+   * A callback that will be called with the incoming message.
+   * Helpful for debugging.
+   */
+  incomingPayload?: IncomingPayloadCallback;
+
+  isStandalone?: boolean;
+};
+
+export type StartServerOptions = Pick<SideCarOptions, "basePath" | "filesToServe" | "incomingPayload"> & {
+  port: number;
+};
