+++ packages/sidecar/src/utils/extras.ts
@@ -1,0 +1,62 @@
+import { get } from "node:http";
+import { startSpan } from "@sentry/node";
+import { SERVER_IDENTIFIER } from "~/constants.js";
+import { logger } from "~/logger.js";
+
+export const withTracing =
+  (fn: CallableFunction, spanArgs = {}) =>
+  (...args: unknown[]) =>
+    startSpan({ name: fn.name, ...spanArgs }, () => fn(...args));
+
+export function logSpotlightUrl(port: number): void {
+  logger.info(`You can open: http://localhost:${port} to see the Spotlight overlay directly`);
+}
+
+export const isValidPort = withTracing(
+  (value: string | number) => {
+    if (typeof value === "string") {
+      const portNumber = Number(value);
+      return /^\d+$/.test(value) && portNumber > 0 && portNumber <= 65535;
+    }
+    return value > 0 && value <= 65535;
+  },
+  { name: "isValidPort", op: "sidecar.server.portCheck" },
+);
+
+export const isSidecarRunning = withTracing(
+  (port: string | number | undefined) =>
+    new Promise(_resolve => {
+      logger.info(`Checking if we are already running on port ${port}`);
+      const options = {
+        hostname: "localhost",
+        port: port,
+        path: "/health",
+        method: "GET",
+        // This is only the socket timeout so set up
+        // a connection timeout below manually
+        timeout: 1000,
+        headers: { Connection: "close" },
+      };
+
+      let timeoutId: NodeJS.Timeout | null = null;
+      const healthReq = get(options, res => {
+        const serverIdentifier = res.headers["x-powered-by"];
+        resolve(serverIdentifier === SERVER_IDENTIFIER);
+      });
+      const destroyHealthReq = () => !healthReq.destroyed && healthReq.destroy(new Error("Request timed out."));
+      function resolve(value: boolean) {
+        process.off("SIGINT", destroyHealthReq);
+        if (timeoutId) {
+          clearTimeout(timeoutId);
+        }
+        _resolve(value);
+      }
+      process.on("SIGINT", destroyHealthReq);
+      timeoutId = setTimeout(destroyHealthReq, 2000);
+      healthReq.on("error", () => {
+        resolve(false);
+      });
+      healthReq.end();
+    }),
+  { name: "isSidecarRunning", op: "sidecar.server.collideCheck" },
+);
