--- packages/sidecar/src/vite-plugin.ts
@@ -1,11 +1,0 @@
-import { setupSidecar } from "./main.js";
-
-export default function spotlightSidecar(port?: number | string) {
-  return {
-    name: "spotlightjs-sidecar",
-
-    configureServer() {
-      setupSidecar({ port });
-    },
-  };
-}
