+++ packages/sidecar/tsconfig.json
@@ -14,5 +14,5 @@
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": ["src/*"]
+  "include": ["src/**/*"]
 }
