+++ packages/sidecar/vitest.config.ts
@@ -1,0 +1,10 @@
+import path from "node:path";
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    alias: {
+      "~": path.resolve(__dirname, "src"),
+    },
+  },
+});
